tools/tx-sim/requirements.txt
autopep8==2.0.1
contextlib2==21.6.0
pycodestyle==2.10.0
PyYAML==6.0
schema==0.7.5
